+++ packages/orq-rc/src/orq_ai_sdk/models/listidentitiesop.py
@@ -208,7 +208,7 @@ class ListIdentitiesData(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-04-27T14:57:59.356Z")
+    updated: Optional[datetime] = parse_datetime("2026-04-27T21:07:34.093Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")