+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KQ7Q5D4G87VYXMDXCYQTQNVQ"
+        "01KQ8CA4JR8GY70RMYS40GY5WV"
     )
     r"""The unique identifier of the data source"""