+++ packages/orq-rc/RELEASES.md
@@ -7359,3 +7359,13 @@ Based on:
 - [python v4.9.0-rc.3] packages/orq-rc
 ### Releases
 - [PyPI v4.9.0-rc.3] https://pypi.org/project/orq-ai-sdk/4.9.0-rc.3 - packages/orq-rc
+
+## 2026-04-28 04:24:04
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.761.9 (2.881.4) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.9.0-rc.4] packages/orq-rc
+### Releases
+- [PyPI v4.9.0-rc.4] https://pypi.org/project/orq-ai-sdk/4.9.0-rc.4 - packages/orq-rc